#65 [refactor] : status로 패키지명을 변경한다
