+++ src/main/java/side/onetime/controller/ScheduleController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import side.onetime.dto.ScheduleDto;
 import side.onetime.global.common.ApiResponse;
-import side.onetime.global.common.constant.SuccessStatus;
+import side.onetime.global.common.status.SuccessStatus;
 import side.onetime.service.ScheduleService;
 
 import java.util.List;
